requirements.txt
easyocr
pillow
pynput
pyautogui